requirements.txt
pandas>=1.5.0
matplotlib>=3.6.0
python-pptx>=0.6.21
reportlab>=3.6.0
python-dotenv>=0.21.0
kaggle>=1.5.12
google-generativeai>=0.3.0
openpyxl>=3.0.0
Pillow>=9.3.0
